+++ CHANGELOG.md
@@ -12,6 +12,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Nothing
 
 
+### Added
+- Checks for issuer & data data in a did:peer JWANT
+### Changed
+- Consolidate DID crypto facilities into "vc" directory
+
 
 ## [3.6.2] - 2024.07.07 - 7fb49766ffd739886d7cc28871cfc3914870fd9d
 ### Added